requirements.txt
dbt-core==1.10.11
dbt-duckdb>=1.8.0